third_party/blink/web_tests/wpt_internal/trust-tokens/trust-token-api-e2e-iframe.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Tests trust token issuance and redemption</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/get-host-info.sub.js"></script>

<body>
    <script>
        'use strict';

        promise_test(() => {
            const frame = document.createElement('iframe');
            frame.src = '/wpt_internal/trust-tokens/resources/trust_token_redemption.py';
            frame.trustToken = JSON.stringify({ type: 'token-redemption' });
            document.body.appendChild(frame);

            return new Promise(resolve => {
                    frame.addEventListener("error", resolve("iframe raised an error"));
                }).then(event => {
                    assert_equals(event, "iframe raised an error");
                });
        }, 'Trust token redemption using iframes fails without prior issuance.');

        promise_test(() => {
            const frame = document.createElement('iframe');
            frame.src = '/wpt_internal/trust-tokens/resources/trust_token_issuance.py';
            frame.trustToken = JSON.stringify({ type: 'token-request' });
            document.body.appendChild(frame);

            return new Promise(resolve => {
                frame.addEventListener("load", resolve);
            }).then(event => {
                assert_equals(frame.contentWindow.document.body.innerText, 'Trust token issuance succeeded.');
            });
        }, 'Trust token issuance succeeds using iframes.');

        promise_test(() => {
            const frame = document.createElement('iframe');
            frame.src = '/wpt_internal/trust-tokens/resources/trust_token_redemption.py';
            frame.trustToken = JSON.stringify({ type: 'token-redemption' });
            document.body.appendChild(frame);

            return new Promise(resolve => {
                frame.addEventListener("load", resolve);
            }).then(event => {
                assert_equals(frame.contentWindow.document.body.innerText, 'Trust token redemption succeeded.');
            });
        }, 'Trust token redemption succeeds using iframes.');

        promise_test(() => {
            const host_info = get_host_info();
            const frame = document.createElement('iframe');
            frame.src = '/wpt_internal/trust-tokens/resources/trust_token_send_redemption_record.py';
            frame.trustToken = JSON.stringify({ type: 'send-redemption-record', issuers: [host_info.ORIGIN] });
            document.body.appendChild(frame);

            return new Promise(resolve => {
                frame.addEventListener("load", resolve);
            }).then(event => {
                assert_equals(frame.contentWindow.document.body.innerText, 'Trust token RR succeeded.');
            });
        }, 'Trust token RR succeeds using iframes.');
    </script>
</body>
